extract view
