+++ WooCommerce/Classes/POS/Presentation/PointOfSaleDashboardView.swift
@@ -295,19 +295,3 @@ private extension PointOfSaleDashboardView {
 }
 
 #endif
-
-struct PointOfSaleSettingsView: View {
-    @Environment(\.dismiss) private var dismiss
-    var body: some View {
-        NavigationView {
-            Text("Settings")
-                .toolbar {
-                    ToolbarItem(placement: .navigationBarTrailing) {
-                        Button("Done") {
-                            dismiss()
-                        }
-                    }
-                }
-        }
-    }
-}
